requirements.txt
pywebview==3.4
b2sdk==1.6.0
wcmatch==8.1.2

# Linux renderer
gobject==0.1.0
PyGObject==3.40.1

# Windows renderer
cefpython3==66.1